requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.1
google-generativeai>=0.6.0
numpy>=1.26.0
scikit-learn>=1.4.0
pypdf>=4.0.0
waitress>=3.0.0
mcp>=1.1.0
requests>=2.32.0


